made vetVisits dir; updated imports; separated add vet visit to separate file
